Move addDeviceModelFromApi and addManufacturerFromApi logic to seedDatabaseFromApi
