+++ src/lib/database/schema.ts
@@ -1,4 +1,3 @@
-import type { DeviceModelV1, Manufacturer } from "@seamapi/types/devicedb"
 import type { HoistedStoreApi } from "zustand-hoist"
 
 import type { StoredDeviceModelV1, StoredManufacturer } from "lib/models.ts"
@@ -11,8 +10,6 @@ export interface DatabaseState {
 }
 
 export interface DatabaseMethods {
-  addManufacturerFromApi: (manufacturer: Manufacturer) => void
-  addDeviceModelFromApi: (device_model: DeviceModelV1) => void
   addManufacturer: (manufacturer: StoredManufacturer) => StoredManufacturer
   addDeviceModel: (device_model: StoredDeviceModelV1) => StoredDeviceModelV1
   setExternalImageProxyEndpoint: (endpoint: string | null) => void
